Don't raise on failed smoke test exit code
Because we don't catch the exception and then the whole script bails out
so we don't end up printing the error log from the process.

Keep raising on --debug just to maintain the behaviour, eg fail the
build if the process fails.
I'm not sure we actually want this behaviour though. As it is we won't
get the output printed out on failed debug builds. So maybe we
shouldn't be skipping the output check? I don't quite get what the
intention is here.