Stringify type hint to avoid circular import
Reproducers:

    python3 -c 'from qutebrowser.extensions import loader'
    python3 -c 'import qutebrowser.commands.command'

Specifically the first on was being called from misc/qutebrowser.spec in
the nightly installer build jobs.

Running the full application still works fine somehow. So it might be
possible to import things in the right order to avoid the loop. But
since this is part of our API we probably don't want to require too much
juggling.