Adjust frozen resource lookup for PyInstaller 6
On MacOS the executable is now in a different folder than the data
files. The path of the python files (which is set by PyInstaller, since
the files are actually embedded in the executable) are under the right
folder. So navigate relative to the current file, as recommended in https://pyinstaller.org/en/stable/runtime-information.html